locale.c: Remove use of nl_langinfo_l()
I was wrong that it was feasible to determine at Configure time if this
was a thread safe implementation or not.  What could be determined was
if it was obviously not thread-safe.  lThe major advantage of using
this, is hence not valid, so remove the code that handled it specially.

perl.h comments are updated